refactor(retrofit/test): moved testSpinnakerHttpException_NewInstance to SpinnakerHttpException class
